performance profiling and testing

+++ include/component.h
@@ -1,15 +1,15 @@
 #pragma once
-
+#include <stdint.h>
 //Component ids are assigned at compile time as they are parsed by the compiler
 //ComponentType::id<Type>()
 class ComponentType
 {
-	static size_t counter;
+	static uint32_t counter;
 public:
 	template<typename T>
-	static size_t id()
+	static uint32_t id()
 	{
-		static size_t _id = counter++;
+		static uint32_t _id = counter++;
 		return _id;
 	}
 };
