+++ src/component.cpp
@@ -1,2 +1,2 @@
 #include "component.h"
-size_t ComponentType::counter = 0;
+uint32_t ComponentType::counter = 0;
